feat: Finalizei o front end

+++ src/components/TaskItem.tsx
@@ -1,78 +1,63 @@
 import React from "react";
 
 import { useState } from "react";
-import { toast } from "sonner";
 
+import { TaskItemProps } from "../lib/Types";
+import {
+  updateTaskStatus,
+  handleTaskItemStatusChange,
+  getTaskStatusClasses,
+} from "../lib/TasksFunctions";
 import { AboutButton, CheckIcon, LoaderIcon, TrashIcon } from "../assets/icons";
 import Button from "./Button";
 import axios from "axios";
 import { useNavigate } from "react-router";
 
-interface Task {
-  id: string;
-  title: string;
-  description?: string;
-  time: "morning" | "afternoon" | "evening";
-  status: "done" | "in_progress" | "not_started";
-}
-
-interface TaskItemProps {
-  task: Task;
-  onDeleteSucess: (taskId: string) => void;
-  handleCheckboxClick: (taskId: string) => void;
-}
-
-const TaskItem = ({
+const TaskItem: React.FC<TaskItemProps> = ({
   task,
   onDeleteSucess,
   handleCheckboxClick,
-}: TaskItemProps) => {
+  type,
+}) => {
   const [deleteIsLoading, setDeleteIsLoading] = useState(false);
 
-  const navigate = useNavigate();
-
-  const handleDeleteClick = async () => {
+  const handleDeleteTask = async () => {
+    if (deleteIsLoading) return;
     setDeleteIsLoading(true);
-
-    try {
-      await axios.delete(`http://localhost:3000/tasks/${task.id}`);
-      onDeleteSucess(task.id);
-      toast.success("Tarefa deletada com sucesso");
-    } catch (error) {
-      toast.error("Erro ao deletar a tarefa");
-    } finally {
-      setDeleteIsLoading(false);
-    }
+    await onDeleteSucess(task.id);
+    setDeleteIsLoading(false);
   };
 
-  const getStatusClasses = () => {
-    if (task.status === "done") {
-      return "bg-brand-primary text-brand-primary";
-    }
-    if (task.status === "in_progress") {
-      return "bg-brand-process text-brand-brand-danger";
-    }
-    if (task.status === "not_started") {
-      return "bg-brand-dark-blue bg-opacity-10 text-dark-blue";
-    }
-  };
+  const navigate = useNavigate();
+  const statusClasses = getTaskStatusClasses(task);
 
-  const handleAboutClick = (task: Task) => {
+  const handleAboutClick = () => {
     navigate(`/task/${task.id}`);
   };
 
   return (
     <div
-      className={`flex bg-opacity-10 items-center mx-6 my-3 gap-2 px-4 py-3 text-sm rounded-lg ${getStatusClasses()} `}
+      className={`flex bg-opacity-10 items-center my-3 gap-2 px-4 py-3 text-sm rounded-lg 
+      ${statusClasses} 
+      ${type === "all" ? "mx-6" : ""} 
+      ${type === "home" ? "mx-0" : ""}
+    `}
     >
       <label
-        className={`transition relative bg-opacity-100 flex h-7 w-7 cursor-pointer items-center justify-center rounded-lg ${getStatusClasses()}`}
+        className={`transition relative bg-opacity-100 flex h-7 w-7 cursor-pointer items-center justify-center rounded-lg ${statusClasses}`}
       >
         <input
           type="checkbox"
           checked={task.status === "done"}
           className="absolute h-full w-full cursor-pointer opacity-0"
-          onChange={() => handleCheckboxClick(task.id)}
+          onChange={() =>
+            handleTaskItemStatusChange(
+              task.id,
+              task,
+              updateTaskStatus,
+              handleCheckboxClick
+            )
+          }
         />
         {task.status === "done" && <CheckIcon />}
         {task.status === "in_progress" && (
@@ -81,20 +66,22 @@ const TaskItem = ({
       </label>
       <div className="flex w-full justify-between">
         {task.title}
-        <a className="flex items-center">
-          <Button
-            color="ghost"
-            onClick={handleDeleteClick}
-            disabled={deleteIsLoading}
-          >
-            {deleteIsLoading ? (
-              <LoaderIcon className="animate-spin text-brand-dark-gray" />
-            ) : (
-              <TrashIcon className="bg-transparent" />
-            )}
-          </Button>
-          <AboutButton onClick={() => handleAboutClick(task)} />
-        </a>
+        <div className="flex items-center">
+          {type !== "home" && (
+            <Button
+              color="ghost"
+              onClick={() => handleDeleteTask()}
+              disabled={deleteIsLoading}
+            >
+              {deleteIsLoading ? (
+                <LoaderIcon className="animate-spin text-brand-dark-gray" />
+              ) : (
+                <TrashIcon className="bg-transparent" />
+              )}
+            </Button>
+          )}
+          <AboutButton onClick={handleAboutClick} />
+        </div>
       </div>
     </div>
   );
